+++ eng/packages/http-client-csharp/generator/Azure.Generator/test/Providers/ResourceDataProviderTests.cs
@@ -1,4 +1,7 @@
-﻿using Azure.Generator.Providers;
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using Azure.Generator.Providers;
 using Azure.Generator.Tests.Common;
 using Azure.Generator.Tests.TestHelpers;
 using Microsoft.Generator.CSharp.Input;
@@ -12,7 +15,7 @@ namespace Azure.Generator.Tests.Providers
     {
         private const string TestClientName = "TestClient";
 
-        [TestCase()]
+        [TestCase]
         public void ValidateResourceDataProviderIsGenerated()
         {
             var resourceModelName = "responseType";