+++ sdk/storage/Azure.Storage.DataMovement.Blobs/perf/Azure.Storage.DataMovement.Blobs.Perf/Scenarios/DownloadDirectory.cs
@@ -46,7 +46,7 @@ namespace Azure.Storage.DataMovement.Blobs.Perf
         public override async Task RunAsync(CancellationToken cancellationToken)
         {
             StorageResource source = BlobResourceProvider.FromContainer(_sourceContainer.Uri);
-            StorageResource destination = LocalFileResourceProvider.FromDirectory(_destinationDirectory);
+            StorageResource destination = LocalFilesStorageResourceProvider.FromDirectory(_destinationDirectory);
 
             await RunAndVerifyTransferAsync(source, destination, cancellationToken);
         }