+++ sdk/storage/Azure.Storage.DataMovement.Blobs/CHANGELOG.md
@@ -13,8 +13,11 @@
     - `BlobStorageResourceOptions.ContentEncoding`
     - `BlobStorageResourceOptions.ContentDisposition`
     - `BlobStorageResourceOptions.CacheControl`
+- Renamed `BlobStorageResourceContainerOptions.BlobDirectoryPrefix` to `BlobPrefix`
 - Changed `BlobContainerClient.StartUploadDirectoryAsync` to `BlobContainerClient.UploadDirectoryAsync` and added a required `waitUntil` parameter.
 - Changed `BlobContainerClient.StartDownloadToDirectoryAsync` to `BlobContainerClient.DownloadToDirectoryAsync` and added a required `waitUntil` parameter.
+- Removed `BlobsStorageResourceProvider()` empty default constructor
+- Changed `BlobsStorageResourceProvider.FromClient` methods to `static` method
 
 ### Bugs Fixed